refactor: 채팅 관련 컴포넌트 폴더 구조 재정리
